backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
numpy==2.2.5
sentence-transformers==3.0.1
supabase==2.5.0
python-dotenv==1.1.0
